==> scripts/0-full-analysis.R <==
# Full analysis in order
source("scripts/1-preprocessing.R")
source("scripts/2-doublet-classification.R")
source("scripts/3-doublet-removal.R")
source("scripts/4-initial-integrated-clustering.R")
source("scripts/5-post-clustering-qc.R")
source("scripts/6-pruning.R")
source("scripts/7-final-integrated-clustering.R")
source("scripts/8-basic-annotation.R")
source("scripts/9-dimplots.R")
source("scripts/10-cluster-markers.R")
source("scripts/11-differential-abundance.R")
source("scripts/12-differential-gene-expression.R")
source("scripts/13-volcano-plots.R")
source("scripts/14-ora.R")
source("scripts/15-diet.R")
